added needed info in signup

+++ internal/handlers/post/SubmissionHandler.go
@@ -72,7 +72,7 @@ func SubmissionHandler(w http.ResponseWriter, r *http.Request) {
 	for _, doc := range returnedProfile {
 		for key, value := range doc {
 			if key == "userId" {
-				userID = value.(int64)
+				userID = int64(value.(float64))
 				break
 			}
 		}
